authentication_service/requirements/base.txt
django==2.0.7
django-environ==0.4.5
psycopg2-binary==2.8.6
git+https://gbozee:abiola2321@gitlab.com/careerlyft-team/careerlyft-shared.git
django-cors-headers==2.4.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
requests==2.18.4
